+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## v1.2.2
 - Add `conditionalRequestGuards()` API for ExpressJS handlers.
+- Add workaround for `PUT` returning 304 from Minio.
 
 ## v1.1.1
 - Add `serverTimings()` API for ExpressJS performance monitoring.